all bugs has been fixed

+++ app/Http/Middleware/CheckUserStatus.php
@@ -3,6 +3,7 @@
 namespace App\Http\Middleware;
 
 use Closure;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -16,17 +17,20 @@ class CheckUserStatus
     public function handle(Request $request, Closure $next): Response
     {
         if (auth()->check() && auth()->user()->company_id != null) {
-            $user = auth()->user()->whereHas('company', function ($query) {
-                $query->where('is_active', false);
-            })->first();
-
+            
+            $user = User::where('id',auth()->user()->id)->where('is_active', 0)->first();
+          
             if ($user) {
-                $request->user()->tokens()->delete();
-                return response()->json([
-                    'status'        => false,
-                    'status_code'   => 401,
-                    'error'         => trans('messages.staff_account_deactivate')
-                ]);
+                
+                if($user->is_staff){
+                    $request->user()->tokens()->delete();
+                    return response()->json([
+                        'status'        => false,
+                        'status_code'   => 401,
+                        'error'         => trans('messages.staff_account_deactivate')
+                    ]);
+                }
+                
             }
         }
